Add basic qmake out of the box usage
Forward headers are now automatically generated when qmake'ing the
project and examples can be disabled using CONFIG+=no_examples.
The forward headers are unconditionally regenerated at each qmake call,
while that is not optimal, it is sufficient most of the time as the
project file is likely not to be modified.