Update CMakeLists.txt to add a 'd' sulffix to the lib built for debug
This command was added to separate the debug and release version, otherise it was need the rename the lib when alternate the compilation between debug and release in the main application.

I have noted that kdchart makes such a separation but I did not find how it is  done there.